feat(gke): Enable HTTPS for GKE service
This commit enables HTTPS for the GKE service by using a Google-managed SSL certificate.

It introduces the following changes:

A ManagedCertificate resource is created to provision and manage the SSL certificate.
An Ingress resource is created to route external traffic to the service and terminate SSL.
The service.yaml is changed from LoadBalancer to NodePort to work with the Ingress.
The Terraform configuration in gke.tf is updated to create a static IP for the Ingress and to apply the new Kubernetes manifests.
The setup-gke.sh script is updated to provide the correct instructions for accessing the service via HTTPS.